Export egl images to an external gl renderer
Register gl-renderer as a plugin and implement apis to
reference and dereference egl images. Idea is to switch
the composition between gl-renderer and external
renderer. External renderer uses egl display, egl surface
and egl images from gl-renderer and renders.

This opens up a lot of possibilities for 3D UI transition
animations. Also to give a 3D composition output involving
different applications. When animation is complete there will
be an option to switch back composition to gl-renderer.

Signed-off-by: Harsha M M <harsha.manjulamallikarjun@in.bosch.com>
3the node will remain in both used and free lists.